fix issue from removing error throw during auth + rm .env.example + added prefixes and ports to .env and docker compose + updated install.md and readme.md + made it impossible to delete root user

+++ app/api/v1/user/[id]/route.ts
@@ -99,6 +99,8 @@ export const DELETE = auth(async (req) => {
     return resourceNotFound();
   if (user.email === auth)
     return invalidRequestWithError("You cannot delete yourself");
+  if (user.email === "filmatrix@sys.admin")
+    return invalidRequestWithError("You cannot delete the root user");
 
   try {
     await prisma.user.delete({
